+++ packages/index/src/filter/filesize-filter.js
@@ -2,21 +2,12 @@ const log = require('@home-gallery/logger')('index.filter.maxFilesize')
 
 const { parseFilesize } = require('../utils')
 
-const createName2Entry = (entries, maxSize) => {
-  return entries
-    .filter(entry => entry.size > maxSize)
-    .reduce((result, entry) => {
-      result[entry.filename] = entry
-      return result
-    }, {})
-}
-
-const isKnown = (name2entry, filename, stat) => {
-  const entry = name2entry[filename]
+const isKnown = (filename2Entry, filename, stat) => {
+  const entry = filename2Entry[filename]
   return entry && entry.size == stat.size
 }
 
-const createFilesizeFilter = (entries, maxFilesize, keepKnownFiles = false) => {
+const createFilesizeFilter = (filename2Entry, maxFilesize, keepKnownFiles = false) => {
   const maxSize = parseFilesize(maxFilesize)
 
   if (maxSize === false) {
@@ -25,15 +16,8 @@ const createFilesizeFilter = (entries, maxFilesize, keepKnownFiles = false) => {
 
   log.info(`Limit files up to ${maxFilesize} (${maxSize} bytes)`)
 
-  if (keepKnownFiles) {
-    const name2entry = createName2Entry(entries, maxSize)
-    return (filename, stat) => {
-      return stat.size <= maxSize || isKnown(name2entry, filename, stat)
-    }
-  } else {
-    return (_, stat) => {
-      return stat.size <= maxSize
-    }
+  return (filename, stat) => {
+    return stat.size <= maxSize || (keepKnownFiles && isKnown(filename2Entry, filename, stat))
   }
 }